requirements.txt
certifi==2018.8.24
chardet==3.0.4
conda==4.3.16
gevent==1.3.6
greenlet==0.4.14
grequests==0.3.0
idna==2.7
imageio==2.3.0
numpy==1.15.1
pexpect==4.6.0
picamera==1.13
Pillow==5.2.0
ptyprocess==0.6.0
pycosat==0.6.3
pygame==1.9.4.dev0
requests==2.19.1
RPi.GPIO==0.6.3
ruamel.yaml==0.15.64
urllib3==1.23
web.py==0.40.dev1
